+++ admin/tabs/general/configs/ConfigPanel.js
@@ -9,6 +9,7 @@ import {creates, hoistCmp} from '@xh/hoist/core';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {restGrid} from '@xh/hoist/desktop/cmp/rest';
 import {Icon} from '@xh/hoist/icon';
+import {regroupDialog} from '../../../regroup/RegroupDialog';
 import {ConfigModel} from './ConfigModel';
 import {differ} from '../../../differ/Differ';
 
@@ -26,7 +27,8 @@ export const configPanel = hoistCmp.factory({
                     });
                 }
             }),
-            differ()
+            differ(),
+            regroupDialog()
         );
     }
 });