+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 * New `GridModel` config `colChooserModel` replaces `enableColChooser` to allow for more flexible
   configuration of the desktop implementation of `colChooser` in both popover and dialog components.
   * Use `colChooserModel: true` to reimplement default behavior.
-  * See documentation on `ColChooserModel` for more information.
+  * See documentation on `GridModel.ColChooserModelConfig` for more information.
 
 ### 🎁 New Features
 
+* Admin Preference and Config panels now provide bulk regrouping actions.
 * Provides new `CustomProvider` for applications that want to use the Persistence API, but
 need to provide their own storage implementation.
 * Added `restoreDefaults` action to default `StoreContextMenu` for `GridModel`.